refactor: split ruby-generator data.js and operators.js
Extract list generators to data-list.js and math/arithmetic generators
to operators-math-gen.js.

Refs #357
